requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==43.0.0
dataclasses-json==0.6.1
debugpy==1.8.5
decorator==4.4.2
executing==2.0.1
faiss-cpu==1.7.4
fpdf==1.7.2
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google_search_results==2.4.2
googleapis-common-protos==1.63.2
greenlet==3.0.1
httplib2==0.22.0
idna==3.4
imageio==2.35.0
imageio-ffmpeg==0.5.1
importlib-metadata==6.8.0
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.0.330
langchain-experimental==0.0.37
langsmith==0.0.57
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.7
mdurl==0.1.2
moviepy==1.0.3
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numexpr==2.8.7
numpy==1.26.1
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.28.1
packaging==23.2
pandas==2.1.2
parso==0.8.4
Pillow==10.1.0
platformdirs==4.2.2
proglog==0.1.10
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.0
psutil==6.0.0
pure_eval==0.2.3
pyarrow==14.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
PyDrive2==1.20.0
pydub==0.25.1
Pygments==2.16.1
pyOpenSSL==24.2.1
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytube==15.0.0
pytz==2023.3.post1

PyYAML==6.0.1
pyzmq==26.1.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.6.0
rpds-py==0.12.0
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SpeechRecognition==3.10.4
SQLAlchemy==2.0.23
stack-data==0.6.3
streamlit==1.37.1
tenacity==8.2.3
tiktoken==0.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.0.7
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.13
yarl==1.9.2
youtube-transcript-api==0.6.1
zipp==3.17.0
